# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "2.0.2",
    "title": "Swagger City for citizens",
    "description": "Showing off City for citizens work",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "servers": [{ "url": "http://185-237-14-12.cloud-xip.com:8082" }],
...
  "paths": {
    "/api/filters": {
      "get": {
        "summary": "Filters",
        "tags": ["Filters"],
        "description": "Filters",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "type": "string",
            "description": "Event city",
            "example": "Kyiv"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string",
            "description": "Event date, time",
            "example": "2023-08-25"
          },
          {
            "name": "eventType",
            "in": "query",
            "type": "string",
            "description": "Event type",
            "example": "Music Event"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "eventCities": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "description": "Event city",
                        "example": "Kyiv"
                      }
                    },
                    "eventDates": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "description": "Event date",
                        "example": "2023-04-15"
                      }
                    },
                    "eventTypes": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "description": "Event type",
                        "example": "Flower Festival"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        }
      }
    },
    "/api/events": {
      "get": {
        "summary": "Events",
        "tags": ["Events"],
        "description": "Events",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "description": "Event id",
                        "example": 1
                      },
                      "idIdentifier": {
                        "type": "number",
                        "description": "Event id",
                        "example": "450e8400-e29b-41d4-a716-446655440000"
                      },
                      "eventTitle": {
                        "type": "string",
                        "description": "Event title",
                        "example": "Виставка квітів \"Тюльпани на Співочому полі\""
                      },
                      "dateTime": {
                        "type": "string",
                        "description": "Event date, time",
                        "example": "2023-04-15T06:30:45.000Z"
                      },
                      "description": {
                        "type": "string",
                        "description": "Event description",
                        "example": "Проводиться приблизно з середини квітня по середину травня, в залежності від погоди. Окрім тюльпанів зазвичай додаються атмосферні тематичні інсталяції по темі, тема щороку нова."
                      },
                      "eventUrl": {
                        "type": "string",
                        "description": "Event url",
                        "example": "https://www.facebook.com/spivochepole"
                      },
                      "eventImage": {
                        "type": "string",
                        "description": "Event image",
                        "example": ""
                      },
                      "eventAddressId": {
                        "type": "number",
                        "description": "Event address id",
                        "example": 1
                      },
                      "locale": {
                        "type": "string",
                        "description": "Event locale",
                        "example": "uk_UA"
                      },
                      "event_address": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "description": "Event address id",
                            "example": 1
                          },
                          "city": {
                            "type": "string",
                            "description": "Event city",
                            "example": "Київ"
                          },
                          "street": {
                            "type": "string",
                            "description": "Event street",
                            "example": "вул. Лаврська, 31"
                          },
                          "notes": {
                            "type": "string",
                            "description": "Event notes",
                            "example": "Співоче поле розташоване на Печерських пагорбах на правому березі річки Дніпро, біля музею-монументу Батьківщина-мати."
                          },
                          "coordinates": {
                            "type": "string",
                            "description": "Event coordinates",
                            "example": "50.4302484,30.4936464"
                          },
                          "locale": {
                            "type": "string",
                            "description": "Event locale",
                            "example": "uk_UA"
                          }
                        }
                      },
                      "eventTypes": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number",
                              "description": "Event id",
                              "example": 1
                            },
                            "eventType": {
                              "type": "string",
                              "description": "Event type",
                              "example": "Квітковий фестиваль"
                            },
                            "locale": {
                              "type": "string",
                              "description": "Event locale",
                              "example": "uk_UA"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        }
      },

      "post": {
        "summary": "Events",
        "tags": ["Events"],
        "description": "Create new event",
        "parameters": [],
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "description": "Create new event",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "locale": {
                    "type": "string",
                    "description": "Event locale",
                    "example": "uk_UA"
                  },
                  "idIdentifier": {
                    "type": "string",
                    "description": "Event idIdentifier",
                    "example": "450e8400-e29b-41d4-a716-446655440000"
                  },
                  "eventTitle": {
                    "type": "string",
                    "description": "Event title",
                    "example": "В гостях у Петровича44"
                  },
                  "date": {
                    "type": "string",
                    "description": "Event date",
                    "example": "2023-04-22"
                  },
                  "time": {
                    "type": "string",
                    "description": "Event time",
                    "example": "11:11"
                  },
                  "description": {
                    "type": "string",
                    "description": "Event description",
                    "example": "test descr"
                  },
                  "eventUrl": {
                    "type": "string",
                    "description": "Event url",
                    "example": "http://halabuda.com"
                  },
                  "city": {
                    "type": "string",
                    "description": "Event city",
                    "example": "Тернопіль"
                  },
                  "street": {
                    "type": "string",
                    "description": "Event street",
                    "example": "Степана Будного, 12"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Event notes",
                    "example": "Біля гаражу Петровича"
                  },
                  "coordinates": {
                    "type": "string",
                    "description": "Event coordinates",
                    "example": "49.538873573784386, 25.583378061974138"
                  },
                  "eventType": {
                    "type": "string",
                    "description": "Event type",
                    "example": "Посиденьки"
                  },
                  "eventImage": {
                    "type": "string",
                    "description": "Event image",
                    "example": "/public/ffff.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "description": "Event id",
                      "example": 26
                    },
                    "locale": {
                      "type": "string",
                      "description": "Event locale",
                      "example": "uk_UA"
                    },
                    "eventTitle": {
                      "type": "string",
                      "description": "Event title",
                      "example": "В гостях у Петровича44"
                    },
                    "dateTime": {
                      "type": "string",
                      "description": "Event date, time",
                      "example": "2023-04-22T08:11:00.000Z"
                    },
                    "description": {
                      "type": "string",
                      "description": "Event description",
                      "example": "test descr"
                    },
                    "eventUrl": {
                      "type": "string",
                      "description": "Event url",
                      "example": "http://halabuda.com"
                    },
                    "eventImage": {
                      "type": "string",
                      "description": "Event image",
                      "example": "/public/ffff.jpg"
                    },
                    "eventAddressId": {
                      "type": "number",
                      "description": "Event address id",
                      "example": 26
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        }
      }
    },
...
    "/api/image": {
      "post": {
        "summary": "Image",
        "tags": ["Image"],
        "description": "Event image",
        "parameters": [],
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "description": "Event image (multipart/form-data)",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventImage": {
                    "type": "file",
                    "format": "binary",
                    "description": "Event image",
                    "example": "event1697920509934.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation, event image created",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized (email or password is wrong)",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        }
      },

      "delete": {
        "summary": "Image",
        "tags": ["Image"],
        "description": "Delete image",
        "parameters": [],
        "security": [{ "Bearer": [] }],
        "requestBody": {
          "description": "Delete image",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventImage": {
                    "type": "string",
                    "description": "Event image",
                    "example": "event1697920509934.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation, event image deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized (email or password is wrong)",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}